[WIP; haven't done js part yet] Increase edit summary limit in the UI
Increase edit summary (and related fields) UI limit to 767 bytes,
and 255 NFC unicode characters. 767 bytes is chosen as largest we can
have while still having covering indexes, 255 unicode characters
is chosen to be the same as before (for scripts that solely use
ASCII), so we can chat with users about what their expectations
are before changing the limit.

Obvious note that unicode characters are a "rough" concept from
a user prespective, given how sometimes a character with diacritics
is "one" character (If there is a precomposed point for it), and
sometimes its multiple characters (base + combining character).
But that's life, I think this meets close enough to most users
expectations, and I don't think whatever complicated logic would
be needed to identify combining characters is particularly worth
it.

Bug: T6715

+++ includes/content/TextContent.php
@@ -65,14 +65,19 @@ class TextContent extends AbstractContent {
 		return $this; # NOTE: this is ok since TextContent are immutable.
 	}
 
-	public function getTextForSummary( $maxlength = 250 ) {
+	public function getTextForSummary( $maxlength = false ) {
 		global $wgContLang;
 
 		$text = $this->getNativeData();
 
-		$truncatedtext = $wgContLang->truncate(
-			preg_replace( "/[\n\r]/", ' ', $text ),
-			max( 0, $maxlength ) );
+		if ( $maxlength === false ) {
+			$truncatedtext = $wgContLang->truncateForEditSummary(
+				preg_replace( "/[\n\r]/", ' ', $text ) );
+		} else {
+			$truncatedtext = $wgContLang->truncate(
+				preg_replace( "/[\n\r]/", ' ', $text ),
+				max( 0, $maxlength ) );
+		}
 
 		return $truncatedtext;
 	}
